requirements.txt
# Core scraping dependencies
scrapy>=2.11.0
requests>=2.31.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Configuration
PyYAML>=6.0.1

# Database
sqlite3  # Built-in, but listed for documentation

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Optional: For advanced features
# streamlit>=1.29.0  # For future UI
# schedule>=1.2.0    # For future scheduling
